refactor: resolve circular dependencies, bind isMutating button states, and optimize performStatusUpdate (X-Tracker #350)
Split reservation index into a concrete leaf service module, pass isMutating to Accept, Reject, and Cancel trigger buttons, and unify trim, API calls, and Sentry captures inside performStatusUpdate.

Ref: https://github.com/Xchange-Taiwan/X-Talent-Tracker/issues/350
