+++ src/components/reservation/AcceptReservationDialog.tsx
@@ -25,12 +25,14 @@ import type { Reservation } from './types';
 interface Props {
   reservation: Reservation;
   className?: string;
+  disabled?: boolean;
   onAccept?: (payload: { id: string; message: string }) => Promise<void> | void;
 }
 
 export default function AcceptReservationDialog({
   reservation,
   className,
+  disabled = false,
   onAccept,
 }: Props) {
   const [open, setOpen] = useState(false);
@@ -85,7 +87,11 @@ export default function AcceptReservationDialog({
   return (
     <Dialog open={open} onOpenChange={onOpenChange}>
       <DialogTrigger asChild>
-        <Button size="sm" className={cn('min-h-9 px-3', className)}>
+        <Button
+          size="sm"
+          className={cn('min-h-9 px-3', className)}
+          disabled={disabled}
+        >
           接受
         </Button>
       </DialogTrigger>
